video calling app/video calling app/requirements.txt
Flask
Flask-SocketIO
opencv-python
numpy
pillow
dlib
numpy
opencv-python
setuptools